refactor(content): migrate from json to markdown
BREAKING CHANGE: Migration from single JSON file for all content, to
single markdown file for each content. It helps to keep data easy
maintainable and to avoid the need to update different content in the
same place.

The JSON also had to update the index for every content,
which was neither maintainable nor evolutionary. In Markdown each file
has its own index property `position` which is used to keep the order
of the content in a list.